requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
distlib==0.3.8
dotenv==0.9.9
fastapi==0.119.0
filelock==3.20.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.1.0
httpx==0.27.2
openai==1.51.0
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
virtualenv==20.25.3
